refactor: database
