+++ src/entity/boards.ts
@@ -1,5 +1,5 @@
 import {Entity, PrimaryGeneratedColumn, Column, OneToMany, } from "typeorm";
-import Columns from './columns.model';
+import Columns from './columns';
 
 @Entity({name: 'boards'})
 export default class Board {
