--- src/migration/1624019791260-test.ts
@@ -1,11 +1,0 @@
-import {MigrationInterface, QueryRunner} from "typeorm";
-
-export class test1624019791260 implements MigrationInterface {
-
-    public async up(queryRunner: QueryRunner): Promise<void> {
-    }
-
-    public async down(queryRunner: QueryRunner): Promise<void> {
-    }
-
-}
