+++ src/resources/tasks/tasks.router.ts
@@ -1,39 +1,37 @@
 import Express, { Router } from 'express';
 import * as tasksService from './tasks.service';
-import Task from './tasks.model';
 
 const router: Express.Router = Router({ mergeParams: true });
 
-router.route('/').get(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
+router.route('/')
+  .get(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
   await tasksService.getAll(req.params["boardId"]).then(tasks => 
-    res.status(200).json(tasks.map(Task.toResponse))
+    res.status(200).json(tasks)
   ).catch(next)
 });
 
-router
-  .route('/:taskId')
+router.route('/:taskId')
   .get(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
     await tasksService.get(req.params["boardId"], req.params["taskId"]).then(task => 
       res.status(200).json(task)
     ).catch(next);
   });
 
-router.route('/').post(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
+router.route('/')
+  .post(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
   await tasksService.create(req.body, req.params['boardId']).then(task => 
-    res.status(201).json(Task.toResponse((task as Task)))
+    res.status(201).json(task)
   ).catch(next)
 });
 
-router
-  .route('/:taskId')
+router.route('/:taskId')
   .put(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
     await tasksService.update(req.params["boardId"], req.params["taskId"], req.body).then(task => 
-      res.status(200).json(Task.toResponse((task as Task)))
+      res.status(200).json(task)
     ).catch(next)
   });
 
-router
-  .route('/:taskId')
+router.route('/:taskId')
   .delete(async (req: Express.Request, res: Express.Response, next: Express.NextFunction) => {
     await tasksService.remove(req.params["boardId"], req.params["taskId"]).then(() => 
       res.status(200).send()
